[FileIO] Move Gmsh from GeoLib/IO to FileIO.
